+++ src/templates/SourceSenderPlugBase.sol
@@ -7,7 +7,7 @@ import "../lib/BytesLib.sol";
 abstract contract SourceSenderPlugBase is PlugBase {
     function _outboundWithSender(
         uint256 gasLimit_,
-        uint256 chainSlug_,
+        uint32 chainSlug_,
         uint256 fees_,
         bytes memory message_
     ) internal {
